sg_inq: Encode whitespaces for '--export'
When using --export we need to encode whitespaces as otherwise
we'll need to quote the string.
The encoding rules are:
- Encode whitespace with underscore
- Encode consecutive whitespaces with a single underscore
- Strip whitespaces from the start and the end of the string
These rules are for compability with the original scsi_id program.
In addition I've implemented another rule, not found in scsi_id:
- Encode non-printable characters with a dot

Additionally I've added the --export conversion to page 0x80
and implemented some safety guards preventing --export to
be called with anything but page 0x80 and 0x83.

Signed-off-by: Hannes Reinecke <hare@suse.de>