55-scsi_sg3_id.rules: skip sg_inq for recent kernels
Recent kernel have an 'inquiry' sysfs attribute, which presents
the raw inquiry data.
So if the 'inquiry' attribute is present the kernel already has
scanned for VPD pages, and we can safely assume that any non-present
VPD page is not supported.
Consequently we don't need to call sg_inq if 'inquiry' is present,
but can rely on the sysfs attributes.

References: bsc#979436

Signed-off-by: Hannes Reinecke <hare@suse.com>