55-scsi-sg3_utils.rules: run sg_inq for SPC devices
As of 2018 kernel 4.15, the kernel doesn't provide VPD pages for "SPC" devices
(SCSI version 0x03, ANSI INCITS 301-1997) in sysfs. It's usually safe to try
though (no counter-example is known), and for scsi_id compatibility, we have
to try. If the kernel ever changes its policy, this rule can perhaps be dropped.