54-before-scsi-sg3_id.rules: new udev rules
These rules are copied from the proposed new rule file 53-storage-hardware.rules
for systemd/udev, adding additional checks to make sure rules aren't
run twice when 53-storage-hardware.rules is added upstream, and dropping
rules that aren't relevant for sg3_utils.

The purpose of this file is to make sure that low-level hardware detection
(in particular, usb_id and ata_id) takes precedence over SCSI inquiry.
This is mainly for compatibility with scsi_id, and based on the observation
that many USB and ATA devices respond with more accurate data to the USB/ATA
device detection than to SCSI inquiry.