fc_wwpn_id: add condition to FC_TARGET_LUN return
If there is no relevant fc_remote_port or fc_host found, there is no need to return the target_lun number.  Returning with no condition causes a FC_TARGET_LUN value to be present in the udev database for devices that are not fibre attached.  Add condition to check.