sg_inq+sg_vpd: merge VPD page processing
This is being done so that ongoing JSON work is
not duplicated for both sg_vpd and sg_inq. This
might lead to small changes in formatting of VPD
(and standard INQUIRY) output. That in turn might
break parsing code expecting the former output.
Hopefully this will push folks who are parsing to
try out JSON output. However JSON output format
may not be that stable, hoping for feedback on
that front.



git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@959 6180dd3e-e324-4e3e-922d-17de1ae2f315