sg3_utils: Fix issue with rescan-scsi-bus.sh removing hard disks
sg_inq command, which is part of the sg3_utils package, has been modified to output
the string "PDT" instead of "Device_type" in commit
d1c58f173918f22723a79d0ab06bcfb60092e75a.
However, rescan-scsi-bus.sh is consuming the output from sq_inq, which has
not been updated to use the modified output of sg_inq. As a result, running
rescan-scsi-bus.sh with the "-r" option is causing hard disks to be removed.
This commit updates rescan-scsi-bus.sh to use the modified output of
sg_inq to avoid this issue.

Signed-off-by: Guruswamy Basavaiah <bguruswamy@vmware.com>