configure.ac: Fix the getrandom() test
Some systems (older versions of Android) have <sys/random.h> but not
getrandom(). Make sure that presence of getrandom() is detected correctly
on these systems.

Fixes: 9ccb556a21f1 ("sg_dd: setup conditional auto rule for getrandom()")
Signed-off-by: Bart Van Assche <bvanassche@acm.org>

Part 2 of this patch:
ci: Build against two versions of the Android NDK

Build against a version of the NDK that does not support getrandom() and
against a version that supports getrandom().

Signed-off-by: Bart Van Assche <bvanassche@acm.org>



git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@1046 6180dd3e-e324-4e3e-922d-17de1ae2f315